+++ noodles-sam/src/writer/record/sequence.rs
@@ -3,17 +3,28 @@ use std::io::{self, Write};
 use super::MISSING;
 use crate::record::Sequence;
 
-pub fn write_sequence<W>(writer: &mut W, sequence: &Sequence) -> io::Result<()>
+pub fn write_sequence<W>(writer: &mut W, read_length: usize, sequence: &Sequence) -> io::Result<()>
 where
     W: Write,
 {
+    // § 1.4.10 "`SEQ`" (2022-08-22): "This field can be a '*' when the sequence is not stored."
     if sequence.is_empty() {
         writer.write_all(&[MISSING])?;
-    } else {
-        for &base in sequence.as_ref() {
-            let n = u8::from(base);
-            writer.write_all(&[n])?;
-        }
+        return Ok(());
+    }
+
+    // § 1.4.10 "`SEQ`" (2022-08-22): "If not a '*', the length of the sequence must equal the sum
+    // of lengths of `M`/`I`/`S`/`=`/`X` operations in `CIGAR`."
+    if sequence.len() != read_length {
+        return Err(io::Error::new(
+            io::ErrorKind::InvalidInput,
+            "read length-sequence length mismatch",
+        ));
+    }
+
+    for &base in sequence.as_ref() {
+        let n = u8::from(base);
+        writer.write_all(&[n])?;
     }
 
     Ok(())
@@ -24,17 +35,25 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_write_sequence() -> io::Result<()> {
+    fn test_write_sequence() -> Result<(), Box<dyn std::error::Error>> {
         use crate::record::sequence::Base;
 
         let mut buf = Vec::new();
-        write_sequence(&mut buf, &Sequence::default())?;
+
+        buf.clear();
+        write_sequence(&mut buf, 0, &Sequence::default())?;
         assert_eq!(buf, b"*");
 
         buf.clear();
         let sequence = Sequence::from(vec![Base::A, Base::C, Base::G, Base::T]);
-        write_sequence(&mut buf, &sequence)?;
+        write_sequence(&mut buf, 4, &sequence)?;
         assert_eq!(buf, b"ACGT");
+
+        buf.clear();
+        assert!(matches!(
+            write_sequence(&mut buf, 0, &sequence),
+            Err(e) if e.kind() == io::ErrorKind::InvalidInput,
+        ));
 
         Ok(())
     }